requirements.txt
cryptography==42.0.5
Flask==2.3.2
gunicorn
python-dotenv==1.0.1
Requests==2.31.0
snowflake_connector_python==3.7.1
snowflake_sqlalchemy==1.5.1
snowflake-connector-python[pandas]
